ai_server/requirements.txt
# --- FastAPI stack ---
fastapi==0.118.0
starlette==0.48.0
uvicorn[standard]==0.37.0
gunicorn==22.0.0

# --- File upload / images ---
python-multipart==0.0.9
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# --- HTTP client & env ---
httpx==0.27.0
python-dotenv==1.0.1
requests>=2.32.3

# --- Pydantic & friends ---
pydantic==2.11.9
pydantic_core==2.33.2
typing_extensions==4.15.0
annotated-types==0.7.0
anyio==4.11.0
h11==0.16.0
idna==3.10
sniffio==1.3.1
click==8.1.8
colorama==0.4.6
typing-inspection==0.4.1

# --- gRPC & Protobuf  ---
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf==6.32.1
googleapis-common-protos==1.70.0

# --- Database & ORM ---
sqlalchemy==2.0.44

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.3.0

# --- ML & NLP ---
torch>=2.9.0
#kobert-transformers>=0.6.0
#transformers>=4.57.0
konlpy>=0.6.0
pandas>=2.3.3
numpy>=2.3.4
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
